// src/app/interfaces/work.ts
export interface Work {
    idWork: any;
    companyName: any;
    typeContract: any;
    position: any;
    specifyPosition: any;
    laboral_activity: any;
    specifyLaboralActivity: any;
    address: any;
    external_number: any;
    interior_number: any;
    suburb: any;
    postalCode: any;
    municipality: any;
    city: any;
    state: any;
    officePhone: any;
    officePhoneExtension: any;
    nowWorkSeniority: any;
    previousWorkSeniority: any;
    creditApplication: any;
}